requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
boto3==1.34.0 
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
reportlab==4.0.9
